Split sourcemap utility functions into separate file
